Revert "CI: Allow dependabot created PR´s to use secrets 2"
This reverts commit 33d6fb40ccab2b7ced6a9f9debc722093c47c661.

using pull_request_target freezes the version of actions to those in the master branch.
So the updates of actions, triggered by dependabot, produce a wrong positive result, since
the test run with master´s version of actions :(

Final solution for now seems to manually re-run the action on dependabot created PR´s.